New additions: - Anisotropic friction component - simulates static and kinetic friction. Uses different friction coefficients along X and Y axis of the contact point. Must have to simulate behavior of ice-skates, sled, ski or simplified version of tires. - Simple friction component - is a simplified version with a single friction coefficient applied equaly in all direction. Usefull to add friction when you can't provide it using collision mesh. For example, Hovercraft uses this component to simulate friction of air-cushion skirt with the ground. - Air drag component - first aerodynamic component added to MMT. Provides velocity dependent air drag, take into account air density, surface area and shape of the object. Shape influence is approximated using drag coefficient: https://en.wikipedia.org/wiki/Drag_coefficient
Hovercraft is now using simple friction component to simulate friction of the skirt. The amount of friction depends on the mode of the engine: the higher hovercraft hovers the less friction is there. Some tweaking is still required but basic behaviour is there. Another addition is air drag. Addition of airdrag placed a limit on vehicle top speed and removes part of the skidding inertia at high velocity.