Refactoring of PSS Visual Debugger component
PSS Visual Debugger component can now be toggles on/off in real-time.
Toggling in On will add Debug Mode Widget which doesn't do much but
makes it more obvious that debugging is on. Rendering of visual debug
elements is disabled when PSSVD is toggled off, but this doesn't stop
components from sending debug information! Disable debug on components
level when it is not necessary anymore.